+++ src/func_800498F8.c
@@ -1,7 +1,7 @@
 #include "common.h"
 
 /* Clears the slot and returns what was in it. */
-s16 func_800498F8(void) {
+s32 func_800498F8(void) {
     u8 *p = D_8009B458;
     s16 v = *(s16 *)(p + 0x4A4);
