+++ src/func_80058DCC.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_80058DCC(void) {
+s32 func_80058DCC(void) {
     return D_8009AFA0;
 }
