+++ src/func_80058E3C.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_80058E3C(s32 arg0) {
+s32 func_80058E3C(s32 arg0) {
     return D_800F2C40[arg0 * 3616 + 0xE0D];
 }
