+++ src/func_8007BF14.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_8007BF14(void) {
+s32 func_8007BF14(void) {
     return D_800938AD;
 }
