+++ src/func_8007BF30.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_8007BF30(void) {
+s32 func_8007BF30(void) {
     return D_800938B2;
 }
