+++ src/func_8007BF40.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_8007BF40(void) {
+s32 func_8007BF40(void) {
     return D_800938B3;
 }
