+++ src/func_80085320.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-s16 func_80085320(void) {
+s32 func_80085320(void) {
     return D_800FE0CC;
 }
